Voting5 V2.0
V1.0的规划满意度结果是错的，因为两个Int相除没得到正确的满意度，V2.0改成两个double相除

+++ Voting5/voting.h
@@ -31,10 +31,10 @@ class Voting
 {
 	private:
 	public:
-		float t[M2C+1][N2C+1];
+		double t[M2C+1][N2C+1];
 		int rank[N2C+1];
 		int M2, N2;
-		Voting(float table[][N2C+1], int ranking[],int m2,int n2)
+		Voting(double table[][N2C+1], int ranking[],int m2,int n2)
 		{
 			M2 = m2;
 			N2 = n2;
